requirements.txt
pandas
numpy
re
requests
bs4
urllib
openbabel
rdkit
openeye
itertools
networkx
network2tikz
matplotlib
Pillow
ctypes
glob
uncertainties
io
scipy
sklearn
urllib
GPy
deepGP



